cumulative bugfixes
FossilOrigin-Name: 5808354e84fda861b7644d4e0607e9a07d67600b

+++ src/headers/spatialite/gg_structs.h
@@ -368,8 +368,8 @@ extern "C"
 /* Memory File Type */
 	char *path;
 	void *buf;
-	uint32_t size;
-	uint32_t offset;
+	uint64_t size;
+	uint64_t offset;
     } gaiaMemFile;
 /** 
  Typedef for Memory File structure
